refactor(remote): remove p2p fetch call from remote
with the remote API maturing, the line that delinates p2p from remote is
point-to-point vs. network-based content resolution. With that in mind
pulling dataset versions is in need of a change. This removes all chances
for resolving content via IPFS, which I think we should add back with
p2p refactoring and better lib-level logic about when to use p2p