feat(oplog): extension interface for sparse-parent logs
logbook was relying on an under-specified property of the opstore.Logstore interface:
some Logstore implementations always fetch the Parent log as part of a Get request.

UserDatasetBranchLog was using this, which blew up when used with a different
implementation. The solution to this is to push the logic needed by logbook.UserDatasetBranchLog
into the oplog package, and implement it using explicit calls to a oplog.Logstore.

In the future I'd like to both work on the Logstore API and write a set of spec tests
that implementations can validate against, much like we've done with dsref.RefResolver.

At the same time I've defined an exention interface that holds these explicit calls and an optional
method logstores can implement to speed up this process. There are situations where this can be
done with a single trip to a database. The extension interface provides this as an upgrade path
without any changes to the logbook or opstore packages