fix(setup): Fix setup, add many unit tests
Due to a recent refactoring, the `qri setup` command was no longer prompting for a username. This caused the username validation to fail. Passing a username via the `--username` flag was also ignored, as the code in lib/setup.go would overwrite it. Fix these problems, add tests to make sure they won't happen again.

Rename the `--peername` flag to `--username` for `qri setup`