fix(p2p): some clean up around `qri peers connect` and `upgradeToQriConnection` (#1489)
* refactor(p2p): don't double report upgrade error

We already log.Debug that there was an upgrade error inside `upgradeToQriConnection`, there is no need to also `log.Error` this message as well. We expect that many of the connections that we will be making will not "speak" qri. This shouldn't be an error.

* fix(`peers connect`): we need to explicitly upgrade a connection when we connect

When we connect directly to a peer using `qri peers connect`, since the command expects the peer's profile as a return value, we need to explicitly upgrade the connection so we know that the profile exists in the Profile store.

When we refactor, we really should be listening on a channel that is monitoring the profile exchange process & only trying to access the peer profile then. This is a temporary solution.