fix(lib): explicitly shutdown `remoteClient` on `inst.Shutdown()`
The remote client we have in our `Instance` might not be the remote client that was created when the instance itself was created, and therefore might not have the same context as the instance. Since we want the instance to fully clean itself up before returning, we must explicitly `Shutdown` the remote client when we `Shutdown` the instance.

This is potentially a more clean pattern: have each part of instance understand how to shut itself down, and return a channel that closes when it is fully cleaned up. Then we can call `Shutdown` on each process and wait for each to be `Done` before considering the instance fully shut down.