fix(resolve): Resolve refs after merging logbook by profileID
Because peers are constructing logbooks with the same profileID, but different initial op values, the old way of merging logs by using their ID (hash of user creation Op) no longer works. The previous commit allows merging logbook data by using the profileID. However, this breaks logbook's ResolveRef function, which was calculating the ref.ProfileID field by looking up the log by the ParentID field. Since divergent logbook data was created, and we merged it using this new method, the ParentID now refers to non-existent data. This can be solved by using the old refstore resolver, and only consulting logbook for the initID of the dataset. Of course, we still want to move off of refstore, so this change is only temporary until we have a proper resolver (such as dscache) in place.