refactor(p2p): adjust `NewQriNode` & tests so that `localResolver` is a passed in param
We may have instances where the node has an opinion about how to resolve a reference locally in a specific way. We should have `localResolver` be a passed in param.

The `NewQriNode` signature change means changing all the calls to `NewQriNode`, most of which occur in tests. Most tests do not require any p2p ref resolution, however, for any test runners or factories  that create nodes to be used in other tests, we have ensured they have a valid local resolver.