refactor(logsync): adjust `put` func to add `dsref.Ref` as a param
In order to be able to add access control to logbook, we need the given `ref` (aka the resource) and the `author.Name` of the subject of the action. Added the ability to add and get the author name to the `p2pClient` and `httpClient`'s requests.

Adjusted the `put` func to add a `ref dsref.Ref`. Previously, we wouldn't know the resource of the action until we read and decode the `r io.Reader` in the request. However, we should really check to make sure that the action the subject is attempted to do is approved before we get that far. Adding `ref` to `put` allows us to do just that.