refactor(api): Combine /get and /body endpoints
Move the functionality of the /body endpoint into /get. The old /body endpoint
still exists, but it's doing most of the same work. Use a single function
`parseGetReqArgs` to parse the relevant values from the http.Request. Introduce
a generic error handler for the API, which will send better http status codes
if it knows the cause of the specific error it is given.