fix(remote): be more lenient with `ResolveRef` error when trying to remove a dataset
When we go to a remove a dataset from a remote, we very well may have already removed identifying information that exists in other systems that the `localResolver.ResolveRef` relies on.

However, even if `ResolveRef` comes back with an error, it still may have enough identifying information for the remote to continue removing the dataset.

In this case, we were coming up against the way that the `repo` resolves the dataset reference. It first checks to see if the reference can be found in the repo, and if so, it asks logbook for the `InitID`. However, at this point in the process, we have already removed the log data for this reference, so we error that the log is not found. However, the reference has been populated. Not fully populated, since we need the InitID to be considered "fully populated", but populated enough that we have what we need to continue removing the dataset.

To be fully honest, this feels a bit strange. It seems like we maybe should either be storing InitIDs in the repo along with the reference, instead of having to add them later, or we should be removing the repo as a place to resolve a reference in the first place. But! At least it's noted now, and we can move forward :)