feat(version): add details reported by "qri version"
since we ask for the output of "qri version" of bug reports, adding more
info to the version command will really help debugging. Specifically,
getting the git commitish hash to print as part of "qri version".

this requires using "make" to build instead of go install and go build
directly, but is well worth it.

All of this is directly inspired by great work over at
github.com/textileio/powergate