feat(save): support custom timestamps on commit
setting the input dataset commit.Timestamp value will now be preserved instead
of being overwritten with the current time. Input timestamp will ALWAYS be
converted to UTC time zone.

This change makes it much easier to generate dataset histories with commit
timestamps that aren't in chronological order, which is perfectly fine
according to spec, but will cause bugs for any code that relies on such an
assumption.