fix(logbook): remove 'stranded' log histories on new dataset creation
if a user starts the save process but abandons it midway with SIGKILL qri's currently
structured to terminate without any cleanup. This means the 'rollback' preocedure in
an errored save never fires, and the dataset name is left taken but not in the user's
list of datasets. This (admittedly hacky) solution checks for stray histories in this
exact state (name and branch logs declared, no version save) and considers them safe
to remove before creating a new replacement log.