fix(dsfs): remove dataset field computing deadlock
the dsfs.computeFieldsFile `Read` and `handleRows` methods were both contending for
the same mutex lock, which would deadlock the program. Fix this by using a TrackedReader
to count bytes instead of local state, which drops the need for taking the lock on calls
to Read. Should yeild a microscopic performance bump as well