test(changes): refactor tests to use testRunner pattern, test Report
reduce boilerplate by creating a test runner for the changes package,
switch to a nil cache for the underlying Stats service, which removes
the need for temp directories in each test.

Adding a number of conveince methods to the runner makes tests more
readable. Add an easy-path test to the Report function for a 20% test
coverage bump.