feat(logbook): add methods for writing transform run ops
Introduce a new type of branch-level operation to logbook that records details
about a transform run. This commit adds two ways a Run-model operation can get
into logbook. The first is via a new *run.State arg to `WriteVersionSave`:

  WriteVersionSave(ctx context.Context, initID string, ds *dataset.Dataset, rs *run.State) error

Passing a non-nil run.State writes two operations that are connected by
referencing the same RunID. When read back out into a dataset log these
operations are combined into a single DatasetLogItem, which has new fields for
describing runs.

The second way to add a Run operation to a branch log is via `WriteTransformRun`

  WriteTransformRun(ctx context.Context, initID string, rs *run.State) error

This is intended for where we'd like to record a run that didn't create a commit
In this case we record a run op, but no commit op. The run op will contain
details about the failure.

All of this needs tests.