refactor(dsref.VersionInfo): merge logbook.DatasetLogItem into VersionInfo
Remove logbook.DatasetLogItem completely by adding `CommitTitle` and
`CommitMessage` fields to `dsref.VersionInfo`. While we're here, add fields that
commit's executed with the `--apply` flag will need to reflect transform run
info.

`logbook.DatasetLogItem` is creating more trouble than it's worth.
While working with @ramfox on qrimatic code it became clear that adding fields
that qrimatic needs fields added to dsref.VersionInfo, and I had been planning
on adding them to logbook.DatasetLogItem. By factoring these into VersionInfo,
we remove a stray data structure that `lib` can return and drop a bunch of
aliasing cruft.

I've also adjusted the documentation of a VersionInfo to reflect that a
VersionInfo can model a failed run as well as a commit (this is important for
upcoming work on adding run operations to logbook). A VersionInfo that describes
a failed commit *must* have a `Path` field equal to the empty string.

One important thing to note here: the `RunDuration` and `RunStatus` fields
are not stored on a `*dataset.Dataset`, which violates the principle of
`VersionInfo` being a "strict subset" of a dataset.