test(run): add SetIDRand to override randomness source for run.NewID
and fix a bunch of tests. Use this new technique in the cmd package to
get consistent runIDs. Keep in mind each call to run.NewID consumes (I think)
16 bytes from the reader, and will panic if the reader errors. So when
overriding SetIDRand, provide a reader that is > 16x the number od calls you
intend to make before another call to SetIDRand