refactor(transform): introduce transform service, retain ref to instance context
commit 5292b64b5831337862cabfdb0eef3f3b7da22927 changed the context that was
being passed to transform.Apply from a context.Background() call within the lib
package to a ctx passed into lib. Within the HTTP package this context is scoped
to an HTTP request, which introduces a premature cancellation bug when Apply is
called asynchonously with `wait = true`.

To fix this, introduce a "service" pattern to the transform package that accepts
a root context, and whenever `wait = true`.

One limitation of this approach: we aren't currently transferring any values
from the passed in context to the background context, meaning downstream code
can't make assumptions it should be able to. In the future we'll need to either
work out a method to transfer important values between contexts (possibly by
re-associating a root context), or refactor the "process context or request
context paradigm, moving the whole conversation farther up the call stack.