refactor(api): middleware translates gorilla.Mux vars to reference string
Keep the "principle" reference string parsing within the lib param by having the
api package map gorilla mux variables to a URL query parameter via middleware.
This keeps the details of how gorilla.Mux variable names map to path components
within the api package itself, and lets lib.GetParams work with ref strings.

Mapping to the `refstr` query param sets up future code that can marshal
directly to a lib.Params struct that has a `Refstr` field, which we should be
trying to harmonize around. Callers of lib should not need to import dsref to
make a lib method call.

There is a downside to overriding the "direct" user-supplied query params: any
response sent back to the client built from the request URL needs to be scrubbed
of the server-side-added "refstr" arg to keep user feedback consistent. I'd
considered embedding this value on `*http.Request.Context()`, but this means
that code downstream of this middleware has to look in more places to build up
request parameters.

In my opionion the pattern of scrubbing responses is a good thing, and could be
expanded in the future to provide the most concise url possible by removing any
unnecessary user-supplied query params, etc.