fix(base.SaveDataset): move logbook write operation back down from lib
This fixes a few tests that were failing thanks to one-off save implmentations
that no longer received logbook updates with logbook write operations moved up
into lib. This was pretty tough to track down

Interaction between subsystems on the save path has become hard to predict. We
have code that implictly require base.SaveDataset manages updates to history
changes. It's test code, but the problem is the implicit-ness, and brittleness
of the codebase

I think th eway forward on both is to use the event bus, and focus on wiring the
bus up properly in all circumstances. If logbook updates from save events
happened via a blocking event bus subscription we could avoid this brittleness
As-is we need to plumb a *run.State argument down into base.SaveDataset just to
pass along to logbook's write.