refactor(dsref): replace ProfileID with InitID in refstring serialization
This is a backwards-compatible spec change to the way we parse dataset reference
strings, specifically the concrete part of a ref (the bit after the `@`). The
grammar change from this:
    <concreteRef> = '@' [ <profileID> ] '/' <network> '/' <commitHash>
to this:
    <concreteRef> = '@' [ <datasetID> | <profileID> ] '/' <network> '/' <commitHash>

The only change is replacing `profileID` with `datasetID`, where `datasetID` is
a lowercase-base32 dataset identifier (initID). This change makes it possible to
reference datasets by ID directly. For example:

    @jzk63pfkfefgwub7wvhl54s4inqr25ncwhmi7svahjjxzwuknrlq
    @jzk63pfkfefgwub7wvhl54s4inqr25ncwhmi7svahjjxzwuknrlq/ipfs/QmHaash

We can make this change backwards-compatible because dataset identifiers &
profileIDs use different character sets. Replacing the profileID is the right
decision, as the use of profileIDs in dsrefs has been inconsitent to date, and
will only get less useful as we transition toward additional collaborators on
the same dataset.