fix(key): fix local keystore key.ID encoding, require ID match keys
Local key store now requires that any private or public key added to the
store *must* match it's key identifier (must be hash of public key, or
inlined public key). By having the keystore require key identifiers
cryptographically match the public key they're storing we have the same
assertion in reverse when we fetch the key by it's itentifier from the
store. This is a crucial proof when key identifiers are used in tokens.

Also fix a number of small errors & discrepencies between local & in-mem
stores of profiles & keys