feat(dispatch): Dispatch func can return 1-3 values, 2 being Cursor
Func implementations registered with Dispatch can return 1-3 values.
Only 1 value means it must be an error. 2 values means the first is
any output and the second is an error. 3 values means the first is any
output, the second is a Cursor, and the third is an error. Cursor will
be used for pagination of methods that support it.