fix(changes): "left" side of report should be the previous path
The upper layers & the UI powered by the `Report` method expect that if the `leftRef` is empty, we should just use the `PreviousPath` of the `rightRef` as the `left` dataset to compare to.

This bug only occurs if the `left` side is empty, and currently all our `ChangeReport` consumers have filled in both versions to compare. There is now a test that ensures the behavior is correct.