refactor(repo): `NewMemRepo` should not create new subsystems if one is provided
Currently, the `NewMemRepo` function creates its own logbook and dscache, with no option for passing an already created cache in, if one was provided.

Since we can have potentially created a separate `logbook` or `dscache` in `lib.NewInstance` that we want to preserve, we must be able to pass these existing subsystems in.