refactor(lib): `newTestRunner` uses `lib.NewInstance`
yay!

The previous way we were creating an instance `NewInstanceFromConfigAndNode` was creating subtle bugs and discrepancies, specially around profile and the concept of the instance `owner`.

This only relies on the `config.DefaultMemConfigForTesting` function, `lib.NewInstance`, and the existence of a `repoPath`. `lib.NewInstance` still expects a `repoPath`, there are some subsystems that in order to work, need to have an on disk place to read and write (as far as I can tell, this is only `Stats` right now, `logbook` can be coerced into writing to the mem fs).

Because the profile/owner of the instance is consistent, we can use `s.ActiveProfile().Peername` as the username resolver for `NewParseResolveLoadFunc`. This allows users to use `me` as a stand-in for their peername.