fix(token): `claim` now includes `ProfileID`
If there are any tokens in a request, we expect that token to contain a profile id that helps the receiving node establish the request's identity. `token.NewPrivKeyAuthToken` was creating a correctly signed token with the priv key and id, but did not include the `ProfileID`. In `inst.activeProfile()` we use `ProfileID` to determine identiy. Because there was no profile id, `profiles.GetProfile` was erroring with "path not found", which is now "profile not found"