refactor(lib): dispatch only POSTs over RPC
Because we don't have knowledge of what information we need to put into
a url string when we dispatch an http rpc call, we still need each
endpoint to use POST w/ application JSON.

This commit also removes the erroneous `{path:.*}` from the end of
endpoints expecting the path to be a ref. Gorilla.mux handles the paths
for us before & we use the `AttributesSet` to determine our endpoints.