refactor(api): use `lib.NewHTTPRequestHandler` to handle dataset requests
list of handlers that were wholesale replaced by `NewHTTPRequestHandler`:
- ListHandler POST GET
- ListRawHandler  POSt GET
- ValidateHandler POST GET
- ManifestHandler POST
- ManifestMissingHandler POST
- DAGInfoHandler POST
- ChangesHandle POST GET
- DiffHandler POST GET
- RenameHandler POST PUT

This means that any tests associated with the above handlers have also been removed.

All other handlers were adjusted to the most current pattern (... that isn't `NewHTTPRequestHandler`)

Known RPC bugs:
Get - rpc doesn't error but no response - The http api is expecting a `GetResponse` over the wire (since that is the return type of the `dataset.get` method). However, our api handler does further adjustments to the `GetResponse` depending on what the user requested in their inital request. Eg, if "meta" is selected, only the meta section is returned. There is pagination if the body was requested, etc. So, unlike most of our endpoints, our api response is not the same as our lib response params. When Dispatch tries to decode the response into a `GetResponse`, it can't because the response isn't shaped like `GetResponse`. We may need to unify the possible get responses & figure out how to represent all variations in a `GetResponse`, and have the `lib.Get` method return only what is needed, rather than returning the entire dataset and having the api and cmd layers narrow down what needs to be shown.