refactor(lib): rename log methods to remove stutter and reduce confusion
We had the concept of `Logs`, as in a list of logs, and `Logbook`, as in the entire log book, switched around in the previous naming. For clarity, I've renamed the endpoints and lib methods:

From `lib.Log` to `lib.History` - this function does not actually return a list of logs, it returns a list of version infos, aka the history rather than the list of log entries

From `lib.Logbook` to `lib.Entries` - this function does not actually return a Logbook, it returns a list of log entries

From `lib.PlainLogs` to `lib.RawLogbook` - this function returns an entire logbook as json. The `cmd` layer referred to this as `Raw`, which feels more like someone who is using lib as a library will grok easier than `Plain`, but this one is definitely up for debate. Regardless if we use `Plain` or `Raw` as a prefix, we should use `Logbook` as the suffix, to distinguish this method as returning the entire logbook (rather than a list of log entries).

The `/logbook` endpoint (which is on the slate for deprecation), should probably return a logbook, rather than a list of log entries.

`/log` now returns a list of log entries, which is where we are headed in api 0.10.0 anyway