refactor(cmd): adjust `render` command to work w/ new render paradigm
Since there is no longer a separation between `RenderViz` and `RenderReadme`, just one `Render` lib method, this commit adjusts the `qri render` command to prepare the params properly for the `Render` method.

There is also a bit of an oddity surrounding the `Remote` params. We had it in the `Readme` renderer and not in the `Viz`. It seems like `scope` is supposed to be handling some of this, but based on our current patterns, we are only able to use `instance.WithScope` at the cmd level. The current `qri render` command does not have a flag that allows the user to specify a remote.

This commit also combines two separate `RenderViz` tests into one