refactor(load): dsref.Loader overhaul
Before, we had the Loader interface with the method LoadDataset, which expected to receive a resolved reference and loaded a dataset. In addition there was the ParseAndResolveLoadFunc type which would parse a ref and resolve it and then load it. This latter behavior more closes matches the `load_dataset` function in starlark.

This change renames the Loader interface's old method to LoadResolved, and moves the ParseAndResolveLoadFunc implementation to a second method on the interface, now named LoadDataset.

Remove the source parameter from scope.ParseAndResolveRef. Instead, the scope is assigned the source at construction time. Many commands require a certain source, so they check for it and return an error upon mismatch.