refactor(load): Simplify Loader more, various cleanups
Loader now only provides LoadDataset. LoadResolved has been removed. For base, cleanup the simpler implementation that it has, and use that for transform and sql tests.

Change handling of dataset names so that it is now an error to use upper-case letters, not just a warning. Rename can still be used to fix bad names

fsi loading is handled by the loader via a useFSI flag, which scope sets for the few commands that require it.

Fix the bug with listWarning that prevented `list` from working if the user's refstore contained bad username / profileID combinations

Remove some TODOs that are no longer applicable.

TODO: lib/load_test is not working due to changes in the semantics of LoadDataset. Need to figure out how to fix it before merging.