feat(automation): `orchestrator` struct and constructor
Adds an `Orchestrator`, a system for managing automation on Qri.
Currenly the orchestrator has a bare-bones `workflow.Store`, takes a
`SaveFactory` and `ApplyFactory`, that direct the orchestrator
on what to run during `RunWorkflow` and `DryRunWorkflow`.

The bare-bones workflow store can only `s.Create` and `s.Get`.

An integration test is up that covers constructing a new orchestrator,
creating a workflow, getting a workflow, running a workflow, and dry
running a workflow.