feat(workflow): add `Trigger` and `Hook` interface
Add a `Trigger` interface that defines what we need for a workflow
trigger to function. Also adds a `TestTrigger` implementation that we
can use to test triggers and workflows.

Add a `Hook` interface that defines what we need for a workflow to
initiate a hook. Also adds a `TestHook` implementation that we can use
to test hooks and workflows.

Also adds triggers and hooks to the `AssertStore` testing function