feat(base): support progress in save, shorter stats keys
upgrade dataset & qfs dependencies, bringing in support for NDJSON & compression
settings in the process. Changes to stats frequency keys have the unfortunate
side effect of breaking hashes & blowing up the size of this commit.

On the qfs side we gain a SizeFile interface, and use it in the save path to show
import progress