test(collection): use EventListener/EventPublisher spec pattern
@dustmop made a good point while reviewing collection code calling for less required
reliance on the event package:

> I think it's better to totally decouple this package [collection] and the interface
> from the event bus.
> Code that does't care about the event bus at all, such as cloud, won't need to
> depend on it by depending on collection.

https://github.com/qri-io/qri/pull/1738/files/a6e9c38530a7729efc6e3e77af05f12362769173#r641002632

I've been thinking about this a bunch, and think we can refine the concern. The event
bus has quickly sprung up to be a highly-relied upon inter-subsystem API, which means
it does need tests, but at the same time the point stands that some of our interface
implementations won't rely on events to maintain state. Instead they'll push that
concern onto a transactional system that manages state transitions like a database.

To that end, I'd like to use two new rules around the intersection of events & spec:

1. Spec tests should split events into two categories: Subscription and Publishing.
2. Event subscription spec must be optional, publishing spec is required.

I totally agree that a subsystem that emits no events shouldn't even need to import
the event package, at the same time, subsystems that do use the event bus need to be
able to rely on all implementations to honor the event publishing contract.

This commit changes the collection spec suite, adding the "Subscription" category
and clarifying that this spec is optional.