fix(run, workflow): fix offset index bug & `Copy` method signature
Previously, there was a bug in listing for both the run.Store and the
workflow.Store that would allow for out-of-bounds indexing of a slice.

Also, the `workflow.Copy` and `run.Copy` signatures have changed from
(using run as an example)

`func (r *State) Copy(run *State)`

to

`func (r *State) Copy() *State`

In the first example, it's unclear which State would be getting copied
from and which would be getting copied into. In the second version, it's
clear that we are returning a copy of the receiver.