feat(trigger): add `Listener`, `Source` interfaces, update `Workflow` methods
The `trigger.Listener` listens for `trigger.Triggers` of a particular
trigger.Type, emitting the `event.ETTriggerWorkflow` event on a given
bus whenever a trigger meets its own specified conditions.

The `trigger.Source` interface abstracts a `workflow.Workflow`, only
giving a `trigger.Listener` what it needs to get whatever information is
necessary to keep a store of up to date triggers.