refactor(trigger): add `ID` method to `Trigger` interface
In order to correctly `Advance` triggers, the `Orchestrator` needs to know which `Trigger` fired. This commit adds an `ID` method to the `Trigger` interface, to lay the groundwork so that the Orchestrator can differentiate between `Triggers` in a `Workflow`.

Since the `trigger.ID` only needs to be unique with respect to the `Workflow.ID` it is associated with, `trigger.NewID()` generates a random a 5 digit string made up of alphanumeric characters.