refactor(orchestrator): adjust `RunWorkflow` path
1) Adjust the `RunWorkflow` path:

When a user manually triggers running a `Workflow`, we do not need to `Advance` any `Triggers`, since no `Trigger` was fired.

We'v separate out the bulk of the running work into an un-exported `runWorkflow`, which is given a `Workflow` to run.

The `RunWorkflow` method still takes a `Context` and a `Workflow.ID`, and is responsible verifying that the `Workflow` exists, before passing that `Workflow` to `runWorkflow`.

We've also adjusted the `handleTrigger` method. It first `Advance`s the `Trigger` that was triggered and saves the updated `Workflow` before calling `runWorkflow`

Possible bug: if a user adjusts the `Workflow` after the `Workflow` has been triggered, but before it has been run we will be in trouble. Might need to have a list of "locked" `Workflows` that we verify before adjusting the `workflow.Store`. Perhaps as part of the theoretical "sync" system.

2) In our first design iteration, we decided that the way we were going to "trigger" manual runs was to have the `RunWorkflow` method send an event over the `event.Bus` that would be caught by `handleTrigger` & would block until the run was complete. This was our initial method for ensuring the `Orchestrator` would only run one `Workflow` at a time. However, since we rely on the `event.Bus` to update the `run.Store` during the run itself, we realized we can't have the `Workflow` run block the `event.Bus`.

We ultimately decided to make manual runs "imperative" and instead have both the manual run and triggered run call `runWorkflow` as described above, and that syncing would be handled by a different "sync" system.

However, the way we created the `Orchestrator` in the `NewOrchestrator` constructor still followed the assumptions of the first design iteration: that we need to handle the `Triggers` when we started regardless of if the `Orchestrator` was "running", because we needed that handler to catch "manual" triggers.

This refactor moves subscribing `handleTriggers` to `o.Start`, which logically is the "start" of running the orchestrator. In a follow-up commit, we will add `unsubscribing` to the `event` package, and unsubscribe `handleTriggers` when we run `o.Stop`